Removed warnings

+++ Game/src/Game.cpp
@@ -1306,11 +1306,11 @@ void Game::UpdateEntityFromMessage(Entity e, message<GameMsg>& msg, bool skip)
 							cText.pos = transform->position;
 						}
 
-						cText.timeRendered = omp_get_wtime();
+						cText.timeRendered = static_cast<float>(omp_get_wtime());
 						cText.pos.y += 15;
 						cText.end_pos = cText.pos;
 						cText.end_pos.y += 50;
-						cText.amount = std::abs(health->currentHealth - hp.currentHealth);
+						cText.amount = std::abs(static_cast<int>(health->currentHealth - hp.currentHealth));
 						GetScene("Game").PushCombatText(cText);
 					}
